+++ src/pages/mypage/activities/[id]/edit/index.tsx
@@ -1,27 +1,14 @@
 import MyPageLayout from '@/components/MyPage/MyPageLayout';
-import { ReactElement, useEffect, useState } from 'react';
-import { FieldValues, useForm } from 'react-hook-form';
+import { ReactElement } from 'react';
 import ActivitiesForm from '@/components/MyPage/Activities/ActivitiesForm';
-import { priceFormat } from '@/utils/priceFormat';
-import { useRouter } from 'next/router';
 import { getActivitiesId } from '@/api/activities';
 import { GetServerSidePropsContext, InferGetServerSidePropsType } from 'next';
-import { QueryClient, dehydrate, useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
-import { PatchMyActivityReq } from '@/types/myActivities';
-import { patchActivitiesId } from '@/api/myActivities';
-import toast from 'react-hot-toast';
-import { AxiosError } from 'axios';
-import { Activity } from '@/types/common/api';
+import { QueryClient, dehydrate } from '@tanstack/react-query';
 import QUERY_KEYS from '@/constants/queryKeys';
-
-interface activityEditMutationProps {
-  activityId: number;
-  data: FieldValues;
-}
+import useMyActivitiesEdit from '@/hooks/Mypage/Activities/Edit/useEdit';
 
 export const getServerSideProps = async (context: GetServerSidePropsContext) => {
   const activityId = Number(context.query['id']);
-
   const queryClient = new QueryClient();
 
   await queryClient.prefetchQuery({
@@ -33,73 +20,7 @@ export const getServerSideProps = async (context: GetServerSidePropsContext) => 
 };
 
 function ActivityEdit({ activityId }: InferGetServerSidePropsType<typeof getServerSideProps>) {
-  const { data: activityData } = useQuery<Activity>({
-    queryKey: [QUERY_KEYS.activityEnroll, activityId],
-    queryFn: () => getActivitiesId(activityId),
-  });
-
-  const [latlng, setLatlng] = useState<{ lat: number; lng: number } | null>(null);
-  const router = useRouter();
-  const queryClient = useQueryClient();
-
-  const methods = useForm<FieldValues>({
-    mode: 'onBlur',
-    defaultValues: {
-      title: activityData?.title || '',
-      price: priceFormat(activityData?.price || 0) || '',
-      address: activityData?.address || '',
-      description: activityData?.description || '',
-      category: activityData?.category || '',
-      bannerImageUrl: activityData?.bannerImageUrl || '',
-      schedules: activityData?.schedules || '',
-
-      subImageUrls: activityData?.subImages || '',
-      subImageIdsToRemove: [],
-      subImageUrlsToAdd: [],
-      scheduleIdsToRemove: [],
-      schedulesToAdd: [],
-    },
-  });
-
-  const activityEditMutation = useMutation({
-    mutationFn: ({ activityId, data }: activityEditMutationProps) =>
-      patchActivitiesId(activityId, data as PatchMyActivityReq),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: [QUERY_KEYS.activityEnroll, activityId] });
-      router.push('/mypage/activities');
-    },
-    onError: (error) => {
-      if (error instanceof AxiosError) toast(`${error.response?.data.message}`);
-    },
-  });
-
-  const handleOnSubmit = async (data: FieldValues) => {
-    delete data.subImageUrls;
-    delete data.schedules;
-    data.price = Number(data.price.replace(/,/g, ''));
-
-    activityEditMutation.mutate({ activityId, data });
-  };
-
-  //처음 불러올때 받은 주소 -> 위도 경도로 바꿔주는 함수
-  const calculateLatlng = async (addressData: string) => {
-    if (addressData) {
-      const response = await fetch(
-        `https://maps.googleapis.com/maps/api/geocode/json?address=${encodeURIComponent(addressData)}&key=${process.env.NEXT_PUBLIC_GOOGLE_MAP_API_KEY!}`,
-      );
-      const data = await response.json();
-      if (data.results && data.results.length > 0) {
-        const location = data.results[0].geometry.location;
-        const newLatlng = { lat: location.lat, lng: location.lng };
-        setLatlng(newLatlng);
-      }
-    }
-  };
-
-  useEffect(() => {
-    calculateLatlng(activityData?.address || '');
-  }, []);
-
+  const { methods, handleOnSubmit, latlng } = useMyActivitiesEdit(activityId);
   return <ActivitiesForm methods={methods} handleOnSubmit={handleOnSubmit} latlng={latlng} isEdit={true} />;
 }
