+++ src/pages/mypage/activities/add/index.tsx
@@ -1,61 +1,10 @@
 import MyPageLayout from '@/components/MyPage/MyPageLayout';
 import { ReactElement } from 'react';
-import { FieldValues, useForm } from 'react-hook-form';
 import ActivitiesForm from '@/components/MyPage/Activities/ActivitiesForm';
-import { postActivities } from '@/api/activities';
-import { PostActivitiesReq } from '@/types/activities';
-import { useRouter } from 'next/router';
-import toast from 'react-hot-toast';
-import { useMutation } from '@tanstack/react-query';
-import { AxiosError } from 'axios';
-
-export interface IsDateTime {
-  date: string;
-  startTime: string;
-  endTime: string;
-}
+import useMyActivitiesAdd from '@/hooks/Mypage/Activities/Add/useMyActivitiesAdd';
 
 function ActivityAdd() {
-  const router = useRouter();
-  const methods = useForm<FieldValues>({
-    mode: 'onBlur',
-    defaultValues: {
-      title: '',
-      price: '',
-      address: '',
-      category: '',
-      description: '',
-      bannerImageUrl: '',
-      schedules: [],
-
-      subImageUrls: [],
-      subImageIdsToRemove: [],
-      subImageUrlsToAdd: [],
-
-      scheduleIdsToRemove: [],
-      schedulesToAdd: [],
-    },
-  });
-
-  const activityAddMutation = useMutation({
-    mutationFn: (data: PostActivitiesReq) => postActivities(data),
-    onSuccess: () => {
-      router.push('/mypage/activities');
-    },
-    onError: (error) => {
-      if (error instanceof AxiosError) toast(`${error.response?.data.message}`);
-    },
-  });
-
-  const handleOnSubmit = async (data: FieldValues) => {
-    delete data.subImageIdsToRemove;
-    delete data.subImageUrlsToAdd;
-    delete data.scheduleIdsToRemove;
-    delete data.schedulesToAdd;
-    data.price = Number(data.price.replace(/,/g, ''));
-    activityAddMutation.mutate(data as PostActivitiesReq);
-  };
-
+  const { methods, handleOnSubmit } = useMyActivitiesAdd();
   return <ActivitiesForm methods={methods} handleOnSubmit={handleOnSubmit} />;
 }
