+++ src/pages/mypage/activities/index.tsx
@@ -1,16 +1,16 @@
 import styles from './Activities.module.css';
-import { ReactElement, useEffect } from 'react';
+import { ReactElement } from 'react';
 import MyPageLayout from '@/components/MyPage/MyPageLayout';
-import useInfiniteScroll from '@/hooks/common/useInfiniteScroll';
 import ActivitiesCard from '@/components/MyPage/Activities/ActivitiesCard';
 import NoDataImg from '#/images/img-noData.png';
 import Image from 'next/image';
 import Link from 'next/link';
-import { QueryClient, dehydrate, useInfiniteQuery } from '@tanstack/react-query';
+import { QueryClient, dehydrate } from '@tanstack/react-query';
 import { getMyActivities } from '@/api/myActivities';
 import { setContext } from '@/api/axiosInstance';
 import { GetServerSidePropsContext } from 'next';
 import QUERY_KEYS from '@/constants/queryKeys';
+import useMActivities from '@/hooks/Mypage/Activities/useMyActivities';
 
 export const getServerSideProps = async (context: GetServerSidePropsContext) => {
   const queryClient = new QueryClient();
@@ -25,24 +25,7 @@ export const getServerSideProps = async (context: GetServerSidePropsContext) => 
 };
 
 function Activities() {
-  const { isVisible, targetRef } = useInfiniteScroll();
-
-  const { data: myActivityItems, fetchNextPage } = useInfiniteQuery({
-    queryKey: [QUERY_KEYS.myActivities],
-    queryFn: ({ pageParam }) => getMyActivities({ cursorId: pageParam }),
-    initialPageParam: 0,
-    getNextPageParam: (lastPage) => {
-      const currentCursorId = lastPage.cursorId;
-      return currentCursorId;
-    },
-  });
-
-  const filteredMyActivities = myActivityItems?.pages.flatMap((page) => page.activities);
-  useEffect(() => {
-    if (isVisible) {
-      fetchNextPage();
-    }
-  }, [isVisible]);
+  const { filteredMyActivities, targetRef } = useMActivities();
   return (
     <>
       <div className={styles.activitiesContainer}>
