+++ src/pages/mypage/reservations/Reservations.module.css
@@ -33,5 +33,4 @@
   color: var(--nomad_black);
   font-weight: 700;
   font-size: 3.2rem;
-  line-height: 5.3rem;
 }
