+++ src/pages/mypage/reservations/index.tsx
@@ -54,7 +54,7 @@ function Reservations() {
   return (
     <div className={styles.container}>
       <div className={styles.header}>
-        <h2 className={styles.label}>예약 내역</h2>
+        <h1 className={styles.label}>예약 내역</h1>
 
         <FilterDropDown type="예약 상태" value={selectedStatus} setValue={setSelectedStatus} />
       </div>
